Iron out all extra compiler warnings
mbim-msg.c:83:17: error: comparison between signed and unsigned integer expressions
cli.c:50:10: error: comparison between signed and unsigned integer expressions
cli.c:81:10: error: comparison between signed and unsigned integer expressions
cli.c:104:10: error: comparison between signed and unsigned integer expressions
cli.c:140:10: error: comparison between signed and unsigned integer expressions
cli.c:154:18: error: comparison between signed and unsigned integer expressions
cli.c:171:10: error: comparison between signed and unsigned integer expressions
cli.c:194:10: error: comparison between signed and unsigned integer expressions
cli.c:223:10: error: comparison between signed and unsigned integer expressions
cli.c:229:17: error: comparison between signed and unsigned integer expressions
cli.c:241:17: error: comparison between signed and unsigned integer expressions
cli.c:247:17: error: comparison between signed and unsigned integer expressions
cli.c:259:17: error: comparison between signed and unsigned integer expressions
cli.c:272:10: error: comparison between signed and unsigned integer expressions
cli.c:549:16: error: comparison between signed and unsigned integer expressions
mbim-dev.c:68:21: error: comparison between signed and unsigned integer expressions
mbim-dev.c:97:10: error: comparison between signed and unsigned integer expressions

+++ mbim-dev.c
@@ -56,7 +56,7 @@ int
 mbim_send(void)
 {
 	struct mbim_message_header *hdr = (struct mbim_message_header *) mbim_buffer;
-	int ret = 0;
+	unsigned int ret = 0;
 
 	if (le32toh(hdr->length) > mbim_bufsize) {
 		fprintf(stderr, "message too big %d\n", le32toh(hdr->length));
@@ -94,7 +94,7 @@ mbim_recv(struct uloop_fd *u, unsigned int events)
 	if (cnt < 0)
 		return;
 
-	if (cnt < sizeof(struct mbim_message_header)) {
+	if (cnt < (ssize_t) sizeof(struct mbim_message_header)) {
 		perror("failed to read() data: ");
 		return;
 	}
